backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.20
python-dotenv==1.0.0
imapclient==2.3.1
pydantic==2.5.0
pyjwt==2.8.0
